refactor(ejs): 替换原先loading图片
